+++ client/src/components/2_organisms/amendEventCard/amendEventCard.tsx
@@ -1,4 +1,4 @@
-import React, { useContext, useEffect } from 'react'
+import React, { useContext } from 'react'
 // Components
 import {
   Icon,
